@@ -1,37 +1,31 @@
 name: wechat_kit
 description: A powerful Flutter plugin allowing developers to auth/share/pay with natvie Android & iOS Wechat SDKs.
-version: 1.1.2
+version: 1.1.3
 # author: v7lin <v7lin@qq.com>
 homepage: https://github.com/v7lin/wechat_kit
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.10.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  convert: ^2.0.2
-  crypto: ^2.0.6
-
-  uuid: '>=1.0.0 <3.0.0'
-
-  json_annotation: '>=2.0.0 <4.0.0'
+  convert: '>=2.0.2 <=3.0.0'
+  crypto: '>=2.0.6 <=3.0.0'
+  uuid: '>=1.0.0 <=3.0.1'
+  json_annotation: '>=2.0.0 <=4.0.0'
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-
-  image: ^2.0.7
-
-  path: ^1.6.4
-  path_provider: ^1.4.0
-
-  okhttp_kit: ^1.0.0
-
+#
+#  image: ^2.0.7
+#  path: ^1.6.4
+#  path_provider: ^1.4.0
+#  okhttp_kit: ^1.0.0
   pedantic:
-
   build_runner:
   json_serializable:
 
