@@ -10,8 +10,8 @@ part 'wechat_auth_resp.g.dart';
 )
 class WechatAuthResp extends WechatSdkResp {
   WechatAuthResp({
-    int errorCode,
-    String errorMsg,
+    int? errorCode,
+    String? errorMsg,
     this.code,
     this.state,
     this.lang,
@@ -21,10 +21,10 @@ class WechatAuthResp extends WechatSdkResp {
   factory WechatAuthResp.fromJson(Map<dynamic, dynamic> json) =>
       _$WechatAuthRespFromJson(json);
 
-  final String code;
-  final String state;
-  final String lang;
-  final String country;
+  final String? code;
+  final String? state;
+  final String? lang;
+  final String? country;
 
   @override
   Map<dynamic, dynamic> toJson() => _$WechatAuthRespToJson(this);
