@@ -9,15 +9,15 @@ part 'wechat_launch_mini_program_resp.g.dart';
 )
 class WechatLaunchMiniProgramResp extends WechatSdkResp {
   WechatLaunchMiniProgramResp({
-    int errorCode,
-    String errorMsg,
+    int? errorCode,
+    String? errorMsg,
     this.extMsg,
   }) : super(errorCode: errorCode, errorMsg: errorMsg);
 
   factory WechatLaunchMiniProgramResp.fromJson(Map<dynamic, dynamic> json) =>
       _$WechatLaunchMiniProgramRespFromJson(json);
 
-  final String extMsg;
+  final String? extMsg;
 
   @override
   Map<dynamic, dynamic> toJson() => _$WechatLaunchMiniProgramRespToJson(this);
