@@ -9,8 +9,8 @@ part 'wechat_subscribe_msg_resp.g.dart';
 )
 class WechatSubscribeMsgResp extends WechatSdkResp {
   WechatSubscribeMsgResp({
-    int errorCode,
-    String errorMsg,
+    int? errorCode,
+    String? errorMsg,
     this.templateId,
     this.scene,
     this.action,
@@ -21,11 +21,11 @@ class WechatSubscribeMsgResp extends WechatSdkResp {
   factory WechatSubscribeMsgResp.fromJson(Map<dynamic, dynamic> json) =>
       _$WechatSubscribeMsgRespFromJson(json);
 
-  final String templateId;
-  final int scene;
-  final String action;
-  final String reserved;
-  final String openId;
+  final String? templateId;
+  final int? scene;
+  final String? action;
+  final String? reserved;
+  final String? openId;
 
   @override
   Map<dynamic, dynamic> toJson() => _$WechatSubscribeMsgRespToJson(this);
