@@ -4,7 +4,7 @@ import 'dart:typed_data';
 
 import 'package:flutter/material.dart';
 import 'package:image/image.dart' as image;
-import 'package:okhttp_kit/okhttp_kit.dart';
+// import 'package:okhttp_kit/okhttp_kit.dart';
 import 'package:path/path.dart' as path;
 import 'package:path_provider/path_provider.dart' as path_provider;
 import 'package:wechat_kit/wechat_kit.dart';
@@ -167,101 +167,101 @@ class _HomeState extends State<Home> {
           ListTile(
             title: const Text('图片分享'),
             onTap: () async {
-              OkHttpClient client = OkHttpClientBuilder().build();
-              Response resp = await client
-                  .newCall(RequestBuilder()
-                      .get()
-                      .url(HttpUrl.parse(
-                          'https://www.baidu.com/img/bd_logo1.png?where=super'))
-                      .build())
-                  .enqueue();
-              if (resp.isSuccessful()) {
-                Directory saveDir = Platform.isAndroid
-                    ? await path_provider.getExternalStorageDirectory()
-                    : await path_provider.getApplicationDocumentsDirectory();
-                File saveFile = File(path.join(saveDir.path, 'timg.png'));
-                if (!saveFile.existsSync()) {
-                  saveFile.createSync(recursive: true);
-                  saveFile.writeAsBytesSync(
-                    await resp.body().bytes(),
-                    flush: true,
-                  );
-                }
-                await _wechat.shareImage(
-                  scene: WechatScene.SESSION,
-                  imageUri: Uri.file(saveFile.path),
-                );
-              }
+              // OkHttpClient client = OkHttpClientBuilder().build();
+              // Response resp = await client
+              //     .newCall(RequestBuilder()
+              //         .get()
+              //         .url(HttpUrl.parse(
+              //             'https://www.baidu.com/img/bd_logo1.png?where=super'))
+              //         .build())
+              //     .enqueue();
+              // if (resp.isSuccessful()) {
+              //   Directory saveDir = Platform.isAndroid
+              //       ? await path_provider.getExternalStorageDirectory()
+              //       : await path_provider.getApplicationDocumentsDirectory();
+              //   File saveFile = File(path.join(saveDir.path, 'timg.png'));
+              //   if (!saveFile.existsSync()) {
+              //     saveFile.createSync(recursive: true);
+              //     saveFile.writeAsBytesSync(
+              //       await resp.body().bytes(),
+              //       flush: true,
+              //     );
+              //   }
+              //   await _wechat.shareImage(
+              //     scene: WechatScene.SESSION,
+              //     imageUri: Uri.file(saveFile.path),
+              //   );
+              // }
             },
           ),
           ListTile(
             title: const Text('文件分享'),
             onTap: () async {
-              OkHttpClient client = OkHttpClientBuilder().build();
-              Response resp = await client
-                  .newCall(RequestBuilder()
-                      .get()
-                      .url(HttpUrl.parse(
-                          'https://www.w3.org/WAI/ER/tests/xhtml/testfiles/resources/pdf/dummy.pdf'))
-                      .build())
-                  .enqueue();
-              if (resp.isSuccessful()) {
-                Directory saveDir = Platform.isAndroid
-                    ? await path_provider.getExternalStorageDirectory()
-                    : await path_provider.getApplicationDocumentsDirectory();
-                File saveFile = File(path.join(saveDir.path, 'demo.pdf'));
-                if (!saveFile.existsSync()) {
-                  saveFile.createSync(recursive: true);
-                  saveFile.writeAsBytesSync(
-                    await resp.body().bytes(),
-                    flush: true,
-                  );
-                }
-                await _wechat.shareFile(
-                  scene: WechatScene.SESSION,
-                  title: '测试文件',
-                  fileUri: Uri.file(saveFile.path),
-                  fileExtension: path.extension(saveFile.path),
-                );
-              }
+              // OkHttpClient client = OkHttpClientBuilder().build();
+              // Response resp = await client
+              //     .newCall(RequestBuilder()
+              //         .get()
+              //         .url(HttpUrl.parse(
+              //             'https://www.w3.org/WAI/ER/tests/xhtml/testfiles/resources/pdf/dummy.pdf'))
+              //         .build())
+              //     .enqueue();
+              // if (resp.isSuccessful()) {
+              //   Directory saveDir = Platform.isAndroid
+              //       ? await path_provider.getExternalStorageDirectory()
+              //       : await path_provider.getApplicationDocumentsDirectory();
+              //   File saveFile = File(path.join(saveDir.path, 'demo.pdf'));
+              //   if (!saveFile.existsSync()) {
+              //     saveFile.createSync(recursive: true);
+              //     saveFile.writeAsBytesSync(
+              //       await resp.body().bytes(),
+              //       flush: true,
+              //     );
+              //   }
+              //   await _wechat.shareFile(
+              //     scene: WechatScene.SESSION,
+              //     title: '测试文件',
+              //     fileUri: Uri.file(saveFile.path),
+              //     fileExtension: path.extension(saveFile.path),
+              //   );
+              // }
             },
           ),
           ListTile(
             title: const Text('Emoji分享'),
             onTap: () async {
-              OkHttpClient client = OkHttpClientBuilder().build();
-              Response resp = await client
-                  .newCall(RequestBuilder()
-                      .get()
-                      .url(HttpUrl.parse(
-                          'https://n.sinaimg.cn/tech/transform/695/w467h228/20191119/bf27-iipztfe9404360.gif'))
-                      .build())
-                  .enqueue();
-              if (resp.isSuccessful()) {
-                Directory saveDir = Platform.isAndroid
-                    ? await path_provider.getExternalStorageDirectory()
-                    : await path_provider.getApplicationDocumentsDirectory();
-                File saveFile = File(path.join(saveDir.path, 'timg.gif'));
-                if (!saveFile.existsSync()) {
-                  saveFile.createSync(recursive: true);
-                  saveFile.writeAsBytesSync(
-                    await resp.body().bytes(),
-                    flush: true,
-                  );
-                }
-                image.Image thumbnail =
-                    image.decodeGif(saveFile.readAsBytesSync());
-                Uint8List thumbData = thumbnail.getBytes();
-                if (thumbData.length > 32 * 1024) {
-                  thumbData = Uint8List.fromList(image.encodeJpg(thumbnail,
-                      quality: 100 * 32 * 1024 ~/ thumbData.length));
-                }
-                await _wechat.shareEmoji(
-                  scene: WechatScene.SESSION,
-                  thumbData: thumbData,
-                  emojiUri: Uri.file(saveFile.path),
-                );
-              }
+              // OkHttpClient client = OkHttpClientBuilder().build();
+              // Response resp = await client
+              //     .newCall(RequestBuilder()
+              //         .get()
+              //         .url(HttpUrl.parse(
+              //             'https://n.sinaimg.cn/tech/transform/695/w467h228/20191119/bf27-iipztfe9404360.gif'))
+              //         .build())
+              //     .enqueue();
+              // if (resp.isSuccessful()) {
+              //   Directory saveDir = Platform.isAndroid
+              //       ? await path_provider.getExternalStorageDirectory()
+              //       : await path_provider.getApplicationDocumentsDirectory();
+              //   File saveFile = File(path.join(saveDir.path, 'timg.gif'));
+              //   if (!saveFile.existsSync()) {
+              //     saveFile.createSync(recursive: true);
+              //     saveFile.writeAsBytesSync(
+              //       await resp.body().bytes(),
+              //       flush: true,
+              //     );
+              //   }
+              //   image.Image thumbnail =
+              //       image.decodeGif(saveFile.readAsBytesSync());
+              //   Uint8List thumbData = thumbnail.getBytes();
+              //   if (thumbData.length > 32 * 1024) {
+              //     thumbData = Uint8List.fromList(image.encodeJpg(thumbnail,
+              //         quality: 100 * 32 * 1024 ~/ thumbData.length));
+              //   }
+              //   await _wechat.shareEmoji(
+              //     scene: WechatScene.SESSION,
+              //     thumbData: thumbData,
+              //     emojiUri: Uri.file(saveFile.path),
+              //   );
+              // }
             },
           ),
           ListTile(
